pyproject.toml
[tool.poetry]
name = "thecoderapi"
version = "0.1.0"
description = ""
authors = ["Luis Antonio Correa Leyva <correaleyval@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.1.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
fastapi = "^0.64.0"
aiofiles = "^0.6.0"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
